+++ src/main/java/br/com/maurigvs/surveyapi/exception/EntityNotFoundException.java
@@ -2,7 +2,7 @@ package br.com.maurigvs.surveyapi.exception;
 
 public abstract class EntityNotFoundException extends BadRequestException {
 
-    public EntityNotFoundException(String entityName, Integer entityId) {
+    public EntityNotFoundException(String entityName, Long entityId) {
         super(entityName + " not found by Id " + entityId);
     }
 }
