+++ src/main/java/br/com/maurigvs/surveyapi/exception/QuestionNotFoundException.java
@@ -4,7 +4,7 @@ import br.com.maurigvs.surveyapi.model.Question;
 
 public class QuestionNotFoundException extends EntityNotFoundException {
 
-    public QuestionNotFoundException(Integer questionId) {
+    public QuestionNotFoundException(Long questionId) {
         super(Question.class.getSimpleName(), questionId);
     }
 }
