+++ src/main/java/br/com/maurigvs/surveyapi/exception/SurveyNotFoundException.java
@@ -4,7 +4,7 @@ import br.com.maurigvs.surveyapi.model.Survey;
 
 public class SurveyNotFoundException extends EntityNotFoundException {
 
-    public SurveyNotFoundException(Integer surveyId) {
+    public SurveyNotFoundException(Long surveyId) {
         super(Survey.class.getSimpleName(), surveyId);
     }
 }
