+++ src/main/java/br/com/maurigvs/surveyapi/repository/AnswerItemRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AnswerItemRepository extends JpaRepository<AnswerItem, Integer> {
+public interface AnswerItemRepository extends JpaRepository<AnswerItem, Long> {
 }
