+++ src/main/java/br/com/maurigvs/surveyapi/repository/SurveyRepository.java
@@ -1,12 +1,11 @@
 package br.com.maurigvs.surveyapi.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import br.com.maurigvs.surveyapi.model.Survey;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SurveyRepository extends JpaRepository<Survey, Integer> {
+public interface SurveyRepository extends JpaRepository<Survey, Long> {
 
     boolean existsByTitle(String title);
 }
